[Wave] Build wave runtime in setup.py (#7)
This PR includes the build of wave runtime in the top-level setup.py.
With the recent removal of torch and hip deps in wave runtime, it became
easier to build wave runtime as part of iree-turbine.

Here we move the setup.py from iree/turbine/kernel/wave/runtime into the
top-level setup.py and update the pyproject.toml build deps accordingly.

---------

Signed-off-by: Harsh Menon <harsh@nod-labs.com>