[Wave] Teach compiler to do masking properly on scaled dims. (#32)
Currently scaled dimensions such as K/32 or K/2 only works on aligned
shape(wrt block_k). In order to add masking/unaligned support for this
we add feature to check that we have indeed scaled bound and try to
scale the bound. This is most useful for handling K dims that are not
power of 2 for Llama3. Main changes are:


1. Moving of `is_scaled_dim` into general_utils.py for reusability.
2. Before setting bounds, we build a map to keep all the "scaled dims",
then we go through the bounds
/dim one by one, and if we find any of the dim is scaled, we apply the
scaling to the bound.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>