disable writing to cache when there are debug_log operations (#99)
Because debug_log operations require extra arguments detailed by the
trace, we either need to disable the cache or save the extra debug arg
info. But since I plan to allow arbitrary closures to eventually be
connected to the debug arg info (for automatic analysis, or launching of
a viewer program, etc), which can not necessarily be serialized, let's
just disable caching when there are debug ops.

Signed-off-by: William G Hatch <william@hatch.uno>