Rename print-ir-after-all compiler option (#200)
The name of the compiler option `print-ir-after-all` suggests that it
will print the IR after every wave pass. However, this option is passed
through to mlir and prints the IR after all mlir passes.
This commit renames this option to `mlir_print_ir_after_all`, which makes
this explicit and now shares the name with the option on mlir.

Signed-off-by: Martin Lücke <martinpaul.luecke@amd.com>
Co-authored-by: Martin Lücke <martinpaul.luecke@amd.com>