[compiler] Apply padding in `materialized_shapes` (#201)
This patch fixes numerical error when the problem shape is unaligned and
not constrained. During `minimize_global_loads` `materialized_shape` is
used to calculate the new optimized access pattern and the issue is
solved by applying padding to the unaligned dimension. We haven't
encountered this issue before because the unaligned shape has always
been part of the constraints.

Follow-up to https://github.com/iree-org/wave/pull/155

---------

Signed-off-by: nithinsubbiah <nithinsubbiah@gmail.com>