NFC improving names and docstrings for clarity (#229)
While reading some code to understand especially the expand_graph pass
better, I decided to make some clarifications. I've improved the
docstring on the pass function to be more clear about what “expansion”
means. The code used the word “reduction” a lot for Iterate nodes, which
is not wrong, but it is confusing when also I've been looking at the
ReduceOp. Renaming to relate to the Iterate op improves clarity. Also
this removes some unused code.

Signed-off-by: William G Hatch <william@hatch.uno>