[Wave] Linearize shared memory accesses (#275)
To promote better register usage, we can linearize shared memory
accesses S.T the generated affine_map for local reads will be 1D which
makes it easier for compiler to find a common subexpression between
different vector.loads.

This helps `ds_read` and `ds_write` to use more of the same base
register and rely more on `offset:` field.

Changes brought from this PR/commit:

1. Introduce a `linearize_shared_access` option in WaveCompile

2. Implement a `linearize_index` util function that takes in
IndexSequence and their strides and return a linearized IndexExpression.

3. Implement a `linearize_shared_mem` in `read_write.py` to reinterpret
N-d memref of shared memory into 1D. This is different than
linearize_memref, because this is static, and since the input memref
doesn't have a dynamic offset, we also do not need to specify offset.

4. Modified `read_write.py`'s `_create_vec_read_write` to generate
linearized reads/writes iff memory is shared and has no mask and is not
a scatter/gather.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>