[Dist.] Standalone multi-device runtime wrapper (#222)
This PR introduces `MultiDeviceLaunchable`, a simplified API for
executing IREE compiled models across multiple devices using
iree-runtime. `MultiDeviceLaunchable` has a similar interface to
`Launchable` but focuses only on the core multi-device functionality.

Motivation to create this class:
The original Launchable class wraps the VM runtime API under the Device
class, as such, adding support for multiple devices in the existing
class will be done separately. MultiDeviceLaunchable is implemented as a
temporary measure and as a demo.

Signed-off-by: Sanket Pandit <sanketp@amd.com>