Introduce normal forms
Original review: https://github.com/ftynse/water/pull/41

Introduce the concept of normal forms. Passes on the Wave dialect follow
the concept of normal forms. A module containing Wave dialect operations
may satisfy conditions of several normal forms defined by the
`WaveNormalFormAttr`. When this is the case, the module (or some its
ancestor), is expected to have the corresponding attribute attached. The
attribute enforces normal form invariants in its verifiers. Passes can
then check whether the attribute is present as a precondition for them
to run, i.e., only apply to IR in certain normal forms. Passes may also
update the attribute to reflect new normal forms after they complete,
which the following passes may rely on. For example, the type inference
pass may set the attribute indicating that all types are fully inferred,
which is a precondition for the lowering pass to run.

This allows us to represent invariants at a finer granularity in the
pass pipeline than op- or dialect-level verifiers, which otherwise would
require introducing new operations or dialects to express the invariants
in their verifiers. This also enables passes to specify pre- and
post-conditions in a way that is visible in the IR.

This can be potentially extended using interfaces, deferring
self-verification to each individual opeation and providing a
registration mechanism for normal forms.

This can be further combined with the transform dialect to provide
"typed" passes where the type of the pass arguments and results is
constructed from normal forms it expects in the input and guarantees on
the output. This, in turn, would allow us to only run the normal form
verifiers when the normal form annotation changes.

---------

Signed-off-by: Alex Zinenko <git@ozinenko.com>