[debugging] location tracking enforcement (#292)
[debugging] location tracking enforcement

tl;dr:

- All ops must now have locations through the TorchFX portion of the wave compiler.
- To get locations at the MLIR level, you now need to give a compile option `drop_debug_info_before_mlir`, which defaults to false.

Details:

This PR adds a location checking and enforcement pass to the compiler, `location_check_pass`.
The pass can be used for diagnostics only (which was its original purpose at the start of this PR journey), to print which operations do or don't have locations at any point in the compiler.
However, it also has an `enforce_locations` argument (default false) to raise an exception if an op is missing a location.

This PR adds this pass, in enforcing mode, at the very end of the Wave compiler pipeline, so from now on we need to have locations.
However, it is disabled when location tracking is generally disabled.
But this PR also turns on variable tracking by default, whereas before it was disabled.

But since locations change the textual representation of MLIR, this also adds the `drop_debug_info_before_mlir` compile option.
When `True`, it drops all locations when converting to MLIR.
This allows us to still verify that we are capturing and propagating locations through the compiler, but still have the location-free printing for MLIR in tests and for exploring MLIR output.
After discussion, it was decided that we should make the default be `True`, so locations are dropped by default.
I think we will probably revisit this default, or under what conditions it defaults to true or false, later after we have proper location support through the iree compiler backend.

Signed-off-by: William G Hatch <william@hatch.uno>