[debugging] emit debug attrs for kernel (#340)
to preserve locations through mlir/llvm compilation, we need to emit a
DISubprogram attribute. This necessitates a few others as well. But with
this, and with some changes to the iree compiler, we can emit source
code locations end-to-end.

Note that these attrs don't yet seem to be supported by the python MLIR
bindings, so this commit uses Attribute.parse() to parse the string
representation of the attributes. We can change that if we upstream
support to the python bindings, but I think it's reasonable to just move
forward with string parsing until then.

---------

Signed-off-by: William G Hatch <william@hatch.uno>