Run external MLIR module through Wave (#308)
This PR introduces 2 new scenarios (both are tested): running an external MLIR module that (1) contains upstream only (lower-level than tensor/linalg) dialects or (2) contains a fully iree-fied IR, which is ready to run without any extra transformations.

For (1), we re-use the logic to create all the wrapper around the main kernel function, and transform manually the uses of memref arguments from the existing module to be compatible with the stream.binding arguments from the wrapper.

For (2), we don't need to do anything, we just forward it to IREE/wave runtime.

---------

Signed-off-by: tyb0807 <sontuan.vu@amd.com>