implement TopkOp (#277)
This implements a TopkOp with a similar shape to the ReduceOp.

The algorithm it uses is similar to
https://github.com/sgl-project/sglang/blob/6337d9057c78fc29b31d15bec892bb9013201f6d/sgl-kernel/csrc/moe/moe_topk_softmax_kernels.cu#L185
-- it does a reduction to find the maximum value and index, then masks
the value at that index and repeats.

It also abstracts the conditional creation in ReduceOp, which is no longer really related to the rest of this PR, but the PR had a long and winding road through review, and changed various times.


---------

Signed-off-by: William G Hatch <william@hatch.uno>