Broader support for Water emission (#371)
This adds broader support for converting wave kernels to water:
- Test conversion of a matmul kernel
- Rework water_emitter structure such that iterate nodes and the trace
overall uses the same function to emit the ops of their respective
subgraph (`_emit_ops_from_graph`)
- Support for Iterate, GetResult, ExtractSlice, and SharedMemoryBarrier
wave nodes
- add ExtractSliceOp to the wave mlir dialect

For full functionality, we still require handling of the index attribute
for MMA nodes. This is currently skipped as it contains `Piecewise`
calls in the indexing expression. Support for this will be added in a
follow-up.

---------

Signed-off-by: Martin Lücke <martin.luecke@amd.com>
Co-authored-by: Tim Gymnich <tim@gymni.ch>