Move wave index attribute representation to DictArrayAttr (#386)
The MMA index sequence represents different indexing expressions for the
different operands. This is modeled using sympy.Piecewise expressions,
to keep to a single indexing expression. This design decision prohibits
conversion of the full expressions to Affine.
For propagation in wave the Piecewise expression is specialized here:
https://github.com/iree-org/wave/blob/07e5c25fc5a94a152418b97609f21f348ca5e6e5/wave_lang/kernel/wave/analysis/index_sequence_analysis.py#L414-418
To enable representation of this in water we use similar specialization
logic and move representation of index sequence from a single dictionary
to an array of dictionaries.
Currently, MMA ops are the only operation with more than a single entry
in this array.

---------

Signed-off-by: Martin Lücke <martin.luecke@amd.com>