[Dynamo] Fix fallback to eager mode in Wave inductor backend (#444)
Fixes #445.

The Dynamo backend registry for the Wave backend currently only supports
the `matmul` operation to be optimized during graph tracing.

However, it lists `torch.ops.aten.mm.default` as the only optimizable
op, and the test for the backend uses `torch.matmul` to generate the FX
graph. As `torch.matmul` does not map top `torch.ops.aten.mm.default`
but instead looks something like `<built-in method matmul of type object
at ...>`, we never actually call the Wave backend and Dynamo silently
falls back to eager mode execution.

This patch fixes the above issue and also fixes another subsequent issue
where Wave fails to compile the kernel when `torch.matmul` is added to
the match list, due to the `schedule` argument being a boolean in
`WaveCompileOptions`, instead of a `SchedulingType`.

Also modifies the backend to fix numerical errors arising due to
differences in how Wave expects the inputs and how they were passed in
by taking a transpose of the RHS input tensor.

---------

Signed-off-by: Vinayak Dev <vinayakdev.sci@gmail.com>