Add support for running custom water pipelines via a transform script (#434)
The transform script is passed as a string to the water subprocess,
where it is parsed and applied after the initial emission of the wave MLIR
dialect.
This enables running custom Transform dialect scripts or pass pipelines
(e.g. type inference) on the Wave MLIR representation produced from a
torch.fx trace.

---------

Signed-off-by: Martin Lücke <martin.luecke@amd.com>