Move compilation logic from wave.py to compile.py (#615)
It had nothing to do in the "main" file to start with. Compiler logic is
heavy and depends on IREE among other things. There is no need to import
all of it merely to construct a kernel without compiling it, e.g., in
templates. Furthermore, this allows us not to import IREE when touching
anything from `wave_lang.kernel` as that imports `wave_lang.kernel.wave`
which used to import IREE because of the compiler and no longer does.

See #611

Signed-off-by: Alex Zinenko <git@ozinenko.com>