[water] remove local imports (#616)
Previous commits successfully untangled the IREE dependency that was
transitively imported with anything from the `wave_lang.kernel` package,
largely due to it importing `wave_lang.kernel.wave` which in turn
contained half of the compilation logic. Move local imports to
file-level imports now. Add an assertion to make sure we will not be
clashing with IREE before attempting to import water.

Signed-off-by: Alex Zinenko <git@ozinenko.com>