[mlir] hide symbols in libWaterPythonCAPI (#641)
Add a cmake configuration to hide all symbols except for the binding
symbols when building libWaterPythonCAPI.so on Linux systems. Otherwise
these clash with symbols loaded from libLLVM.so by ROCm 7 that
unconditionally loads libLLVM.so whenever torch is loaded. Arguably it
shouldn't and it's a clutch, but we need to make progress ourselves
without waiting for ROCm to figure it out.

Signed-off-by: Alex Zinenko <git@ozinenko.com>