Decompose memrefs early (#647)
Decompose memrefs early, materializing index calculations, llvm GEPs and
llvm loads/stores explicitly. This makes index calculations more
amendable for integer range and (future) uniformity analyses. Add an
integer narrowing pass to the pipleine to see the reduced count of VGPRs
on mxfp gemm.

Some operations with non-trivial lowering (buffer casts and
`amdgpu.gather-to-lds`) are kept in memref land but converted to 0D
memrefs, to expose indexing as well.

This pass is basically an alternative memref-to-llvm-lowering and it is
compatible with the existing upstream one as for operations not lowered
it will generate `MemrefDescriptor` shims.

* What's wrong with the default upstream memref-to-llvm lowering?
Default lowering uses `MemrefDescriptors`, hiding all index
calculations.
* Why not use 1D/0D memrefs? I want `* sizeof(T)` part of index
calculations to be materialized explicitly. Early POC was using
`memref<?xi8>` and `memref.view` but it required a lot of memref casts
back and forth.
* Why not `ptr` dialect? I had a `ptr` dialect POC as well but `ptr`
dialect is currently incomplete and also at this point it is a carbon
copy of `llvm` dialect and provides no useful abstraction.

---------

Signed-off-by: Ivan Butygin <ivan.butygin@gmail.com>